requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pytest==8.3.5
pytest-cov==6.0.0
pydantic-settings==2.7.1
mysql-connector-python==9.2.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
httpx==0.28.1
